pyproject.toml
[project]
name = "automobil"
version = "0.1.0"
description = ""
authors = [
    {name = "Ricardo Santos",email = "ricardotenv@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic (>=2.11.5,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-ai-slim[mcp] (>=0.2.16,<0.3.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "openai (>=1.84.0,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "mcp[cli] (>=1.9.3,<2.0.0)",
]

[project.scripts]
populate = "automobil.cli.commands:populate"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
